requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
aiomysql == 0.2.0
python-dotenv == 1.0.0
